If it is not possible to load the window state (for example, it is the first time that the application runs), set a default value (33%) for the sidebar width.
This fixes Gitlab issue #3 (bugzilla #788621) which was causing the
sidebar to extend to the whole window (sidebar_width = 0, i.e. 100%),
thus hiding the main panel and rendering the application basically
unusable when the sidebar was enabled.

The second version of this patch introduced further fixes to the
window resizing code (the sidebar panel is correctly resized upon
window resizing or window minimization/maximization).

This third version includes support for the correct migration from
the old window state format to the new format (sidebar width as a
percentage instead of pixels).

Signed-off-by: Guido Trentalancia <guido@trentalancia.com>
---
 src/gdict-window.c |   78 ++++++++++++++++++++++++++++++++++++-----------------
 src/gdict-window.h |    2 -
 2 files changed, 54 insertions(+), 26 deletions(-)